app/requirements.txt
sanic
sanic_compress
jinja2
tinydb
peewee
github3.py